refactor(kbar): extract Crow SSE streaming logic into useCrowChat hook
Moves all state management, SSE parsing, and event handling out of
CrowFallback into a dedicated useCrowChat hook. CrowFallback is now
UI-only. No behaviour change.
